Changing Insert<> method internal usage of int to long to match the actual method signature

+++ Dapper.Contrib/SqlMapperExtensions.cs
@@ -376,7 +376,7 @@ namespace Dapper.Contrib.Extensions
                     sbParameterList.Append(", ");
             }
 
-            int returnVal;
+            long returnVal;
             var wasClosed = connection.State == ConnectionState.Closed;
             if (wasClosed) connection.Open();
